First tage of the fully automatic OBT branches testing sequence generation.
New 'obtSequence.inc' file introduces, to replace same block in settings.sh.

Modify obtSequencer.sh to use obtSequence.inc file if it is exists. (If not
then it try to use the origianl way embedded into settings.sh.)

Add script to generate the 'obtSequence.inc' from the 'master' and the selected
(maxAutoBranch=4) number of the last *.x branches.

Signed-off-by: AttilaVamos <attila.vamos@gmail.com>